requirements.txt
requests~=2.31.0
pandas~=2.0.3
matplotlib~=3.7.2
calmap~=0.0.11
pillow~=10.2.0
tweepy~=4.14.0
python-dotenv~=0.21.0